requirements.txt
absl-py==1.4.0
gdown==4.7.1
lightning-utilities==0.8.0
matplotlib==3.7.1
midi2audio==0.1.1
mido-fix==1.2.12
music21==8.3.0
musicpy
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
Pillow==9.5.0
protobuf==3.19.4
psutil==5.9.5
pygame==2.4.0
pyparsing==3.0.9
pytorch-lightning==2.0.2
pytorch-warmup==0.1.1
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
sf2-loader
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
triton==2.0.0
typing_extensions==4.5.0
wandb==0.15.2
webcolors==1.13
yarl==1.9.2
